data_processing/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
python-dateutil>=2.8.2
uuid>=1.30
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0 